보행 관련 요소 제거

+++ core/util/src/main/java/com/choidev/latesteffort/core/util/motion/MotionManager.kt
@@ -13,7 +13,7 @@ interface MotionManager {
 
     fun observeStep(
         sensorRate: SensorRate = SensorRate.NORMAL,
-        stepEvent: (step: Int) -> Unit
+        stepEvent: (step: Float) -> Unit
     )
 
     fun unregisterObservers()
